isolate test_line_hover and bypass version check
